requirements.txt
Django==4.2.1
djangorestframework==3.14.0
psycopg2==2.9.1
Pillow==8.3.1